Make collect and codemod subcommands of a single executable
