--- bin/prop_requiredness/aggregate.dart
@@ -1,2 +1,0 @@
-export 'package:over_react_codemod/src/prop_requiredness/bin/aggregate.dart'
-    show main;
