--- bin/prop_requiredness/collect.dart
@@ -1,2 +1,0 @@
-export 'package:over_react_codemod/src/prop_requiredness/bin/collect.dart'
-    show main;
